requirements.txt
discord.py
aiohttp
asyncio
python-dotenv
